+++ CHANGES.rst
@@ -37,6 +37,7 @@ fixes:
 - chore: bump mr-smithers-excellent/docker-build-push version (#1633)
 - docs: fix example code in the testing section (#1643)
 - chore: update all core dependencies (#1651)
+- fix: use template file for webserver plugin echo output (#1654)
 
 
 v6.1.9 (2022-06-11)